pyproject.toml
[project]
name = "kitaku"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
